remove class struct method aliases again
We'd need to replicate them in all subclasses as they depend on the class they are called on.
Instead of having duplicated docs it's imo better to document that class struct methods take classes as arguments.
Or as an alternative implement a metaclass hierarchy mirroring the object hierarchy which provides the class stuct methods.