rework search index; include multiple namespace versions. See #76
Deviate from the original sphinx search index structure and add namespaces.
While at it remove namespaces from the filenames and references as
we can concat them at runtime (this saves ~20% in index size).
The behaviour should be the same as before: only the newest
version gets included in the search.