Handling current-module-prefix stripping for stub generation
The problem here is that we need class/struct/etc. references within the
module being annotated to not have the module prefix, because otherwise
we just get unresolvable typing references. We also need all of these
to be forward references (i.e., strings) unless we want to do some kind
of dependency ordering, which I don't.

I'm not a fan of doing this with a global, but the other solutions are
much more complicated. Maybe we can make a stubber state class instead
of this, but it's not obvious how much better that would be.