Handle indirect GI module dependencies for imports
While there's dependency information in the namespace class, it's not
reliable when we need to have all relevant modules in the current
namespace for name resolution. The solution here is to constantly update
a list of modules that have been referenced by current-module
annotations, since that's what mypy, etc. are going to need.

This change also means that we can't write import order until the end of
the stubbing, so we're now keeping the stubs in a StringIO.