Ignore spurious nodes in private member check
The check for whether a record is private assumes a fixed child node
structure that isn't the case in some environments. This patch tries
to keep the spirit of the check by filtering out all known-unimportant
child nodes and simply checking whether there's anything left.

The expected unimportant nodes are:

 * empty text nodes containing formatting white-space
 * source-position elements that map the member to the original source
   file

neither of which indicate that the member actually needs to be
annotated.