docobj: when caching things include the source namespace in the cache key
gio e.g. references some gdk types which shouldn't resolve because gdk isn't
a depdendency. In case we parsed gdk before gio we would parse the docs in the gdk context
and the types would resolve and the resulting class docs would be cached.

Prevent this by adding the source namespace to the cache key.

This should help us get more reproducible output.