namespace: ignore nested records
In Gsk-4.0 the record Gsk.PathPoint contains a nested record.
In case the tag didn't have a name we would just go up the tree
until we found one. In case of a nested record this results in
Gsk.PathPoint being found twice, the second time with the nested
record which should be record.

Fix by not going up for "record" which should never be needed there.